Move and expose SystemInfo
