+++ descopesdk/src/main/java/com/descope/internal/http/DescopeClient.kt
@@ -1,5 +1,6 @@
 package com.descope.internal.http
 
+import com.descope.android.SystemInfo
 import com.descope.sdk.DescopeConfig
 import com.descope.sdk.DescopeSdk
 import com.descope.types.DeliveryMethod
