+++ descopesdk/src/main/java/com/descope/sdk/Sdk.kt
@@ -5,8 +5,8 @@ package com.descope.sdk
 import android.content.Context
 import android.content.pm.ApplicationInfo
 import android.os.Looper
+import com.descope.android.DescopeSystemInfo
 import com.descope.internal.http.DescopeClient
-import com.descope.internal.http.DescopeSystemInfo
 import com.descope.internal.others.ConsoleLogger
 import com.descope.internal.routes.Auth
 import com.descope.internal.routes.EnchantedLink
